添加材质工厂，TypeValue需要一个Unity返回接口

+++ Assets/Editor/Effector/GradientColorEffector.cs
@@ -7,7 +7,7 @@ using Newtonsoft.Json.Linq;
 public class GradientColorEffector : IEffector
 {
     private List<uint> colors = new List<uint>();
-    private List<uint> alphas = new List<uint>();
+    private List<float> alphas = new List<float>();
     private List<float> colorRatios = new List<float>();
     private List<float> alphaRatios = new List<float>();
 
@@ -16,7 +16,7 @@ public class GradientColorEffector : IEffector
     public void deserialize(JObject data)
     {
         colors = StringUtil.SplitString<uint>((string)data["colors"], new char[]{','});
-        alphas = StringUtil.SplitString<uint>((string)data["alphas"], new char[] { ',' });
+        alphas = StringUtil.SplitString<float>((string)data["alphas"], new char[] { ',' });
 
         colorRatios = StringUtil.SplitString<float>((string)data["colorRatios"], new char[] { ',' });
         alphaRatios = StringUtil.SplitString<float>((string)data["alphaRatios"], new char[] { ',' });
